+++ src/basic/components/product/ProductCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Product } from "../../types";
+import { Product } from "../../../types";
 
 interface ProductCardProps {
   product: Product;
